ddsd

+++ src/main/java/sigiv/Backend/sigiv/Backend/dto/detalleVenta/DetalleVentaRequestDto.java
@@ -5,11 +5,13 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 public class DetalleVentaRequestDto {
     private Long productoId;
-    private Integer cantidad;
+    private BigDecimal cantidad;
 }
